flask_serve/requirements.txt
boto3
Flask-WTF
Flask-Bootstrap4
numpy>=1.22.3
matplotlib>=3.5.2
pyyaml
